requirements.txt
# Core packages for data processing
requests>=2.31.0
pandas>=2.1.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Terminal and progress visualization
colorama>=0.4.6
tqdm>=4.66.0

# Data validation and processing
jsonschema>=4.19.0
fuzzywuzzy[speedup]>=0.18.0

# Optional AI processing
openai>=1.3.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.0.0
mypy>=1.5.0

# Logging and monitoring
structlog>=23.1.0